Handling for streaming non-mp3 files
Current behavior hard-codes the stream.mp3 request handler, which does not work for non-mp3 source files unless the server is already set to transcode to mp3. Neither the local nor vlc picroft audio backends support container detection, so the generalized stream legacy handler does not work. This commit changes the request URI to use the current swagger spec, which supports transcoding requests in the URI args; it is hardcoded to mp3 because picroft's local audio backend has a limited number of containers supported.

Since current swagger spec requires user_id in the stream handler and since the skill appears to be using both user auth and token auth, the token was replaced with the user_id to minimize things that can go wrong.